Add new property for hiding/showing comment tabs move entry editor preferences
