+++ src/main/java/org/jabref/preferences/EntryEditorPreferences.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.entryeditor;
+package org.jabref.preferences;
 
 import java.util.Map;
 import java.util.Set;
@@ -48,6 +48,7 @@ public class EntryEditorPreferences {
     private final BooleanProperty autoLinkFiles;
     private final ObjectProperty<JournalPopupEnabled> enablementStatus;
     private final BooleanProperty shouldShowSciteTab;
+    private final BooleanProperty showUserCommentsFields;
 
     public EntryEditorPreferences(Map<String, Set<Field>> entryEditorTabList,
                                   Map<String, Set<Field>> defaultEntryEditorTabList,
@@ -60,7 +61,8 @@ public class EntryEditorPreferences {
                                   double dividerPosition,
                                   boolean autolinkFilesEnabled,
                                   JournalPopupEnabled journalPopupEnabled,
-                                  boolean showSciteTab) {
+                                  boolean showSciteTab,
+                                  boolean showUserCommentsFields) {
 
         this.entryEditorTabList = new SimpleMapProperty<>(FXCollections.observableMap(entryEditorTabList));
         this.defaultEntryEditorTabList = new SimpleMapProperty<>(FXCollections.observableMap(defaultEntryEditorTabList));
@@ -74,6 +76,7 @@ public class EntryEditorPreferences {
         this.autoLinkFiles = new SimpleBooleanProperty(autolinkFilesEnabled);
         this.enablementStatus = new SimpleObjectProperty<>(journalPopupEnabled);
         this.shouldShowSciteTab = new SimpleBooleanProperty(showSciteTab);
+        this.showUserCommentsFields = new SimpleBooleanProperty(showUserCommentsFields);
     }
 
     public ObservableMap<String, Set<Field>> getEntryEditorTabs() {
@@ -210,5 +213,16 @@ public class EntryEditorPreferences {
 
     public void setShouldShowSciteTab(boolean shouldShowSciteTab) {
         this.shouldShowSciteTab.set(shouldShowSciteTab);
+    }
+
+    public boolean shouldShowUserCommentsFields() {
+        return showUserCommentsFields.get();
+    }
+    public BooleanProperty showUserCommentsFieldsProperty() {
+        return showUserCommentsFields;
+    }
+
+    public void setShowUserCommentsFields(boolean showUserCommentsFields) {
+        this.showUserCommentsFields.set(showUserCommentsFields);
     }
 }
